+++ src/main/java/com/um/apitest/pizza/PizzaRepository.java
@@ -2,5 +2,5 @@ package com.um.apitest.pizza;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PizzaRepository extends JpaRepository<Pizza, Long> {
+public interface PizzaRepository extends JpaRepository<Pizza, Integer> {
 }
